requirements.txt
click==8.0.0
cmake==3.18.4.post1
colorama==0.4.4
dlib==19.22.0
face-recognition==1.4.0
face-recognition-models==0.3.0
MouseInfo==0.1.3
numpy==1.20.3
opencv-python==4.5.2.52
pillow==8.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
